+++ app/views/admin/questions/new.html.erb
@@ -1,4 +1,4 @@
-<%= form_for @question do |f| %>
+<%= form_for [:admin, @question] do |f| %>
   <%= f.label :query %>
   <%= f.text_area :query %>
