+++ spec/features/instructor_creates_an_answer_spec.rb
@@ -6,7 +6,7 @@ feature "Instructor Creates an Answer" do
 
   scenario "user adds an answer" do
     pending "Going to pick up on this feature later."
-    visit new_question_answer_path(question)
+    visit new_admin_question_answer_path(question)
     fill_in "Content", with: "SQL"
     click_on "Create Answer"
     expect(page).to have_content("New answer created")
